+++ python/chronos/script/benchmark-chronos.py
@@ -14,154 +14,47 @@
 # limitations under the License.
 #
 
-
-import psutil
-import platform
-import subprocess
-import os
 import sys
-import logging
-import warnings
-from typing import Union, Dict, List, Optional
+import getopt
 import numpy as np
-import re
 
-from utils import get_bytesize, _find_path, get_nano_env_var
-from bigdl.nano.common.common import _env_variable_is_set, _find_library
-
+from utils import CPU_info, check_nano, test_run
+from bigdl.nano.utils.log4Error import invalidInputError
 
 
-def CPU_info():
-    """
-    Capture hardware information, such as CPU model, CPU informations, memory status
-    """
+if __name__ == "__main__":
+    # available options
+    model_list = ["LSTM", "Seq2Seq", "TCN", "Autoformer"]
+    stage_list = ["latency"]
+    opt_list = []
 
-    #information about CPU
-    socket_num = int(subprocess.getoutput('cat /proc/cpuinfo | grep "physical id" | sort -u | wc -l'))
-    model_name = subprocess.getoutput('lscpu | grep "Model name"')
-    model_name = model_name.partition(":")[2]
+    opts, args = getopt.getopt(sys.argv[1:], "m:s:", ["model=", "stage="])
+    for opt, arg in opts:
+        opt_list.append(opt)
+        if opt in ['-m', '--model']:
+            model_name = arg
+        elif opt in ['-s', '--stage']:
+            stage = arg
 
-    print(">"*20,"Hardware Information",">"*20)
-    print('CPU architecture:',platform.processor())
-    print('CPU model name:',model_name.lstrip())
-    print('Logical Core(s):',psutil.cpu_count()) 
-    print('Physical Core(s):',psutil.cpu_count(logical=False)) 
-    print('Physical Core(s) per socket:',int(psutil.cpu_count(logical=False)/socket_num))
-    print('Socket(s):',socket_num)
-    print('CPU usage:',str(psutil.cpu_percent()) + '%')
-    print('CPU MHz:',format(psutil.cpu_freq().current,'.2f'))
-    print('CPU max MHz:',format(psutil.cpu_freq().max,'.2f'))
-    print('CPU min MHz:',format(psutil.cpu_freq().min,'.2f'))
-    print('Total memory:',get_bytesize(psutil.virtual_memory().total))
-    print('Available memory:',get_bytesize(psutil.virtual_memory().available))
+    invalidInputError('-m' in opt_list or '--model' in opt_list,
+                      "-m/--model option must be provided but get none.")
+    invalidInputError(model_name in model_list, f"model argument should be one of {model_list},"
+                      f" but get {model_name}.")
+    invalidInputError('-s' in opt_list or '--stage' in opt_list,
+                      "-s/--stage option must be provided but get none.")
+    invalidInputError(stage in stage_list, f"stage argument should be one of {stage_list},"
+                      f" but get {stage}.")
 
-    #support instruction set or not
-    disabled_logo = "\033[0;31m\u2718\033[0m"
-    abled_logo = "\033[0;32m\u2714\033[0m"
+    latency1, latency2 = test_run(model_name, stage, lookback=96, horizon=48)
 
-    for flag in ["avx512f", "avx512_bf16", "avx512_vnni"]:
-        flag_enabled = int(subprocess.getoutput(f'lscpu | grep -c {flag} '))
-        if flag_enabled:
-            print("Support", flag, ":", abled_logo)
-        else:
-            print("Support", flag, ":", disabled_logo)
-        
-    print("<"*20, "Hardware Information", "<"*20, "\n")
-    
-
-
-
-def check_nano(use_malloc: str = "tc", use_openmp: bool = True) -> None:
-    """
-    Check whether necessary environment variables are setted properly
-    """
-    # Get a copy of os environment
-    env_copy = os.environ.copy()
-    # Get the proper environment
-    correct_env = get_nano_env_var()
-  
-    # Flags about the environment values are proper or not
-    flag = {"LD_PRELOAD" : 1, "tcmalloc" : 1, "Intel OpenMp" : 1, "TF" : 1}
-
-    #Output information
-    name = {"LD_PRELOAD" : "", "tcmalloc" : "", "Intel OpenMp" : ": ", "TF" : ": "}
-    output_list = []
-
-    # Find conda directory
-    conda_dir = None
-    try:
-        conda_dir = subprocess.check_output("conda info | awk '/active env location/'"
-                                            "| sed 's/.*:.//g'",
-                                            shell=True).splitlines()[0].decode("utf-8")
-        conda_env_name = conda_dir.split("/")[-1]
-    except subprocess.CalledProcessError:
-        warnings.warn("Conda is not found on your computer.")
-
-    conda_lib_dir = conda_dir + "/lib" if conda_dir is not None else None
-    openmp_lib_dir = _find_library("libiomp5.so", conda_lib_dir)
-    jemalloc_lib_dir = _find_library("libjemalloc.so", conda_lib_dir)
-    tc_malloc_lib_dir = _find_library("libtcmalloc.so", conda_lib_dir)
-
-    # Detect Intel OpenMP library
-    if use_openmp:
-        if openmp_lib_dir is not None:
-            # Check environment variables
-            for var in ["OMP_NUM_THREADS", "KMP_AFFINITY", "KMP_BLOCKTIME"]:
-                if not _env_variable_is_set(var, env_copy) or env_copy[var] != correct_env[var]:
-                    flag["Intel OpenMp"] = 0
-                    name["Intel OpenMp"] = name["Intel OpenMp"] + var + " "
-                    output_list.append("export " + var + "=" + correct_env[var])
-        else:
-            output_list.append("Intel OpenMP library (libiomp5.so) is not found.")
-
-    # Detect jemalloc library
-    if use_malloc is "je":
-        if jemalloc_lib_dir is not None:
-            if not _env_variable_is_set("MALLOC_CONF", env_copy) or env_copy["MALLOC_CONF"] != correct_env["MALLOC_CONF"]:
-                output_list.append("export MALLOC_CONF=oversize_threshold:1,background_thread:true,metadata_thp:auto,dirty_decay_ms:-1,muzzy_decay_ms:-1")            
-        else:
-            output_list.append("jemalloc library (libjemalloc.so) is not found.")
-
-    if use_malloc is "tc":
-        if tc_malloc_lib_dir is None:
-            flag["tcmalloc"] = 0
-            output_list.append("tcmalloc library (libtcmalloc.so) is not found.")
-    
-    # Check TF_support
-    for var in ["TF_ENABLE_ONEDNN_OPTS", "ENABLE_TF_OPTS", "NANO_TF_INTER_OP"]:
-        if not _env_variable_is_set(var, env_copy) or env_copy[var] != correct_env[var]:
-            flag["TF"] = 0
-            name["TF"] = name["TF"] + var + " "
-            output_list.append("export " + var + "=" + correct_env[var])
-
-    # Check LD_PRELOAD    
-    if not _env_variable_is_set("LD_PRELOAD", env_copy) or not _find_path(env_copy["LD_PRELOAD"]):
-        flag["LD_PRELOAD"] = 0
-        output_list.append("export LD_PRELOAD=" + correct_env["LD_PRELOAD"])
-   
-    # Output overview
-    print(">"*20,"Environment Variables",">"*20)
-    disabled_logo = "\033[0;31mnot enabled \033[0m" + "\033[0;31m\u2718\033[0m"
-    abled_logo = "\033[0;32m enabled \033[0m" + "\033[0;32m\u2714\033[0m"
-    for category in ["LD_PRELOAD", "tcmalloc", "Intel OpenMp", "TF"]:
-        if flag[category]  == 0:
-            print(category, name[category], disabled_logo)
-        else:
-            print(category, abled_logo)
-    
-    # Output suggestions
-    
-    if output_list != []:
-        print(" ")
-        print("+" * 20,"Suggested change: ", "+" * 20)
-        for info in output_list:
-            print(info)
-        print("+" * 60, "\n")
-         
-    print("<"*20, "Environment Variables", "<"*20, "\n")
-
-
-
-if __name__ == '__main__':
+    # output information
+    print("-"*50, "short summary", "-"*50, '\n')
     CPU_info()
     check_nano()
+    print(">" * 20, "nyc_taxi test-run result", ">" * 20)
+    for info in ["p50", "p90", "p95", "p99"]:
+        print(info, "latency:", latency1[info], "ms")
+    if model_name != "Autoformer":
+        for info in ["p50", "p90", "p95", "p99"]:
+            print(info, "latency with onnx:", latency2[info], "ms")
+    print("<" * 20, "nyc_taxi test-run result", "<" * 20)
